Move policy logic into separate file
