+++ src/torchcodec/samplers/__init__.py
@@ -1,4 +1,4 @@
-from ._implem import (
+from ._implementation import (
     clips_at_random_indices,
     clips_at_random_timestamps,
     clips_at_regular_indices,
